+++ lib/boring_backup/stores.rb
@@ -1,3 +1,2 @@
 require_relative "stores/store"
 require_relative "stores/result"
-require_relative "stores/s3" # TODO: load conditionally