+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Configurable storage classes for s3
+
 ## [0.3.0] - 2026-07-13
 
 - Make s3 upload stream part size and thread count configurable